Revert PMF changes.

+++ src/model/expression.rs
@@ -5,7 +5,7 @@ use bio::stats::logprobs;
 use model;
 
 
-pub type PMF = model::pmf::PMF<f64>;
+pub type PMF = model::pmf::PMF<u32>;
 
 
 pub fn pmf(count: u32, count_exact: u32, readout_model: &model::Readout) -> PMF {
@@ -16,7 +16,7 @@ pub fn pmf(count: u32, count_exact: u32, readout_model: &model::Readout) -> PMF 
 
     // TODO trim
     PMF::new(
-        likelihoods.iter().enumerate().map(|(x, lh)| ((offset + x as u32) as f64, lh - marginal)).filter(|&(_, p)| p >= model::MIN_PROB).collect_vec()
+        likelihoods.iter().enumerate().map(|(x, lh)| (offset + x as u32, lh - marginal)).filter(|&(_, p)| p >= model::MIN_PROB).collect_vec()
     )
 }
